requirements.txt
fastapi
lyricsgenius
pydantic
pyyaml
uvicorn